Logs error responses from chain_verifier CheckProposal requests
